chore: /libディレクトリにロジックを移す
