+++ src/Article.ts
@@ -1,5 +1,5 @@
 import { Tags } from './Tags';
-import dayjs from './dayjs';
+import dayjs from './lib/dayjs';
 import { Post } from './post';
 
 export const Article = (post: Post): string => `
