+++ src/post.ts
@@ -1,4 +1,4 @@
-import dayjs from './dayjs';
+import dayjs from './lib/dayjs';
 
 export type Post = {
   id: number;
